Conflicts:
      src/Controller/RegionRestController.php
      src/Controller/SearchRestController.php
      src/Lib/Session.php
      src/Lib/Xhr/XhrMethods.php
      src/Modules/Basket/BasketXhr.php
      src/Modules/Core/View.php
      src/Modules/FoodSharePoint/FoodSharePointControl.php
      src/Modules/Foodsaver/FoodsaverControl.php
      src/Modules/Foodsaver/FoodsaverXhr.php
      src/Modules/Group/GroupTransactions.php
      src/Modules/Message/MessageGateway.php
      src/Modules/Message/MessageTransactions.php
      src/Modules/Profile/ProfileView.php
      src/Modules/Region/RegionControl.php
      src/Modules/Region/RegionXhr.php
      src/Modules/Settings/SettingsControl.php
      src/Modules/Settings/SettingsView.php
      src/Modules/Store/StoreControl.php
      src/Modules/Store/StoreGateway.php
      src/Modules/Store/StoreTransactions.php
      src/Modules/Store/StoreView.php
      src/Modules/StoreUser/StoreUserControl.php
      src/Modules/StoreUser/StoreUserView.php
      src/Modules/WorkGroup/WorkGroupControl.php
      src/Modules/WorkGroup/WorkGroupGateway.php

+++ src/Lib/View/Utils.php
@@ -1264,8 +1264,8 @@ class Utils
 		$id = $this->identificationHelper->id($id);
 		$label = $this->translationHelper->s($id);
 
-		$yearRangeFrom = (isset($option['yearRangeFrom'])) ? $option['yearRangeFrom'] : (date('Y') - 60);
-		$yearRangeTo = (isset($option['yearRangeTo'])) ? $option['yearRangeTo'] : (date('Y') + 60);
+		$yearRangeFrom = (isset($option['yearRangeFrom'])) ? $option['yearRangeFrom'] : ((int)date('Y') - 60);
+		$yearRangeTo = (isset($option['yearRangeTo'])) ? $option['yearRangeTo'] : ((int)date('Y') + 60);
 
 		$value = $this->dataHelper->getValue($id);
